+++ app/src/main/AndroidManifest.xml
@@ -12,12 +12,12 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
             tools:ignore="GoogleAppIndexingWarning">
-        <activity
-                android:name=".Main2Activity"
-                android:label="@string/title_activity_main2"
-                android:theme="@style/AppTheme.NoActionBar">
-        </activity>
-        <activity android:name=".view.activities.MainActivity">
+        <!--<activity-->
+                <!--android:name=".Main2Activity"-->
+                <!--android:label="@string/title_activity_main2"-->
+                <!--android:theme="@style/AppTheme.NoActionBar">-->
+        <!--</activity>-->
+        <activity android:name=".view.activities.ContactListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <action android:name="android.intent.action.VIEW"/>
@@ -26,7 +26,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".view.activities.AddContactActivity"/>
-        <activity android:name=".view.activities.ContactListActivity"/>
     </application>
 
 </manifest>
