+++ app/src/main/res/layout/activity_contact_list.xml
@@ -6,12 +6,24 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:orientation="vertical"
         tools:context=".view.activities.ContactListActivity">
+
+    <Button
+            android:id="@+id/add_contato_pageTXT"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/add_contact"u
+            android:layout_gravity="center"
+            android:textStyle="bold"
+            android:padding="10dp"/>
 
     <android.support.v7.widget.RecyclerView
             android:id="@+id/contato_list_recyclerview"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" >
+            android:layout_height="wrap_content" >
     </android.support.v7.widget.RecyclerView>
+
+
 
 </LinearLayout>
