--- app/src/main/res/layout/activity_main.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:orientation="vertical"
-        tools:context=".view.activities.MainActivity">
-
-    <Button
-            android:id="@+id/add_contato_pageTXT"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Adicionar Contato"
-            android:textStyle="bold"
-            android:padding="30dp"/>
-
-    <Button
-            android:id="@+id/ver_contatosTXT"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Ver Contatos"
-            android:textStyle="bold"
-            android:layout_marginTop="10dp"
-            android:padding="30dp"/>
-
-</LinearLayout>
